% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.R
\name{bootstrap.get.job.runtime.dist}
\alias{bootstrap.get.job.runtime.dist}
\title{Get distribution of job runtime via bootstrap re-sampling}
\usage{
bootstrap.get.job.runtime.dist(size.reps.table, num.bootstrap.reps, runtimes)
}
\arguments{
\item{size.reps.table}{Data frame with 2 columns; typically obtained as the
output from the aggregate() function.
1st column is the task size
2nd column is the number of tasks with this size}

\item{runtimes}{}
}
\description{
Get distribution of job runtime via bootstrap re-sampling
}
\examples{
data(m3xlarge.runtimes.expdist)
job <- c(1,60,100)
srt <- aggregate(job, by=list(job), length)
dist <- bootstrap.get.job.runtime.dist(srt, 500, m3xlarge.runtimes.expdist)
}

