% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulated-annealing.R
\name{compare.schedules}
\alias{compare.schedules}
\title{Compare 2 schedules based on their score}
\usage{
compare.schedules(cur.schedule, proposed.schedule, runtimes, runtimes.summary,
  deadline, max.temp, max.iter, cur.iter)
}
\arguments{
\item{cur.schedule}{Current assigment with score attribute (list)}

\item{proposed.schedule}{Proposed schedule with no score (list)}

\item{runtimes}{Matrix of runtimes for the given instance type. Each row in
the matrix represents a single training sample and has 2 columns. The size
column is the size of task that was processed. The runtime_sec column is the
time taken to process the task in seconds.}

\item{runtimes.summary}{Numeric matrix containing mean and variance of
runtimes for each size}

\item{deadline}{Time by which job must be complete (float). Same time units
as runtimes}

\item{max.temp}{Max temperature to use in the simulated annealing process
'(integer)}

\item{max.iter}{Max # iterations to use to find the optimal schedule via
simulated annealing (integer)}

\item{cur.iter}{Value of current iteration (integer)}
}
\value{
A list containing the accepted schedule and score
}
\description{
Scores are calculated for both schedules. If the score of the proposed
schedule is lower than the score for the current schedule, the proposed
schedule and score are returned. If the score of the proposed schedule is
greater than or equal to the current schedule, the a function of the
current temperature and the 2 scores is used to determine which schedule
to return.
}

