% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.r
\name{get.initial.assignment}
\alias{get.initial.assignment}
\title{Get initial assignment of jobs to instances in a cluster}
\usage{
get.initial.assignment(cluster.size, task.sizes)
}
\arguments{
\item{cluster.size}{Number of instances in the cluster (+ve integer)}

\item{task.sizes}{Array of task sizes (positive reals)}
}
\value{
List containing a mapping of tasks to instances in cluster. The list index represents the id of an instance in the cluster while the associated list member represents the task assigned to that instance
}
\description{
Get initial assignment of jobs to instances in a cluster
}
\examples{
init <- get.initial.assignment(10, seq(1:30))
}

