% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulated-annealing.R
\name{get.initial.schedule}
\alias{get.initial.schedule}
\title{Get initial schedule of jobs to instances in a cluster}
\usage{
get.initial.schedule(cluster.size, task.sizes, runtimes.summary,
  method = "random")
}
\arguments{
\item{cluster.size}{Number of instances in the cluster (+ve integer)}

\item{task.sizes}{Array of task sizes (+ve reals)}

\item{runtimes.summary}{Numeric matrix containing mean and variance of
runtimes for each size. Must be supplied when method='leptf'}

\item{method}{Method to use to assign tasks to instances.
Must be one of ('random', 'leptf').}
}
\value{
List containing a mapping of tasks to instances in cluster.
The list index represents the id of an instance in the cluster while
the associated list member represents the task assigned to that instance
}
\description{
Get initial schedule of jobs to instances in a cluster
}
\examples{
a <- get.initial.schedule(3, 1:30)
rs <- matrix(nrow=2, ncol=3)
rs[1,1] <- 10; rs[1,2] <- 23.5; rs[1,3] <- 2.5
rs[2,1] <- 20; rs[2,2] <- 33.5; rs[2,3] <- 3.5
a <- get.initial.schedule(3, c(rep(10, 3), rep(20, 3)), rs, method='leptf')
}

