% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.r
\name{get.neighbor.by.moving.tasks}
\alias{get.neighbor.by.moving.tasks}
\title{Generate neighbor by moving 1 task}
\usage{
get.neighbor.by.moving.tasks(assignment, num.tasks)
}
\arguments{
\item{assignment}{A list representing the assignment for which a neighbor is desired}

\item{num.tasks}{Integer representing the number of tasks to be moved from 1 instance to another}
}
\value{
A list representing the neighboring assignment
}
\description{
Randomly select 2 instances in the cluster. Randomly select a task from one of the instances and move it to the other instance. Simple random sampling without replacement is used in both sampling stages.
}
\examples{
assignment <- get.initial.assignment(10, seq(1:30))
neighbor <- get.neighbor.by.moving.tasks(assignment, 1)
}

