% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulated-annealing.R
\name{get.neighbor}
\alias{get.neighbor}
\title{Generate a neighbor to an schedule}
\usage{
get.neighbor(schedule)
}
\arguments{
\item{schedule}{A list representing a mapping of tasks to instances in a
cluster}
}
\value{
A list representing the modified schedule of tasks to instances in
the cluster
}
\description{
The input schedule is modified in one of several different ways, including
\itemize{
 \item Move a task from 1 instance to another
 \item Exchange a task with another instance
 \item Move 2 tasks from 1 instance to another
 \item Exchange 2 tasks with another instance
 \item Move 2 tasks from an instance to 2 other instance
 \item Exchange 2 tasks with 2 other instances
 \item and so on...
}
Only the first 2 methods are currently implemented with an equal probability
of selecting either method.
}
\examples{
schedule <- get.initial.schedule(3, 1:30)
proposed.schedule <- get.neighbor(schedule)
}

