% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.r
\name{get.neighbor.single.transfer}
\alias{get.neighbor.single.transfer}
\title{Generate neighbor by moving 1 task}
\usage{
get.neighbor.single.transfer(assignment)
}
\arguments{
\item{assignment}{A list representing the assignment for which a neighbor is desired}
}
\value{
A list representing the neighboring assignment
}
\description{
Randomly select 2 instances in the cluster. Randomly select a task from one of the instances and move it to the other instance. Simple random sampling without replacement is used in both sampling stages.
}
\examples{
assignment <- get.initial.assignment(10, seq(1:30))
neighbor <- get.neighbor.single.transfer(assignment)
}

