% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.r
\name{get.runtimes}
\alias{get.runtimes}
\title{Get training set runtimes for instance type}
\usage{
get.runtimes(instance.type)
}
\arguments{
\item{instance.type}{Instance type of cluster (string). All instances in the cluster are assumed to be of the same type}
}
\value{
Matrix of runtimes for the given instance type. Each row in the matrix represents a single training sample and has 2 columns. The size column is the size of task that was processed. The runtime_sec column is the time taken to process the task in seconds.
}
\description{
This function is called for its side-effect
}
\examples{
get.runtimes('m3xlarge')
}

