% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.r
\name{get.score}
\alias{get.score}
\title{Get score for input assignment}
\usage{
get.score(assignment, runtimes, runtimes.summary, deadline)
}
\arguments{
\item{assignment}{The assignment which needs to be scored (list)}

\item{runtimes}{Training set runtimes (matrix)}

\item{runtimes.summary}{Summary of training set runtimes (matrix)}

\item{deadline}{Time by which job must complete (float; same units as runtimes)}
}
\value{
The score of the assignment, i.e, the probability of the assignment completing the job by the deadline based on the training set runtimes of the tasks in the job (float).
}
\description{
Get score for input assignment
}
\examples{
assignment <- get.initial.assignment(2, c(1,60,100,100,100,230,400,800,2000,2700))
runtimes <- matrix(c(1,48,2555), nrow=1, ncol=3)
prob <- get.score(assignment, runtimes, 600)
}

