% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.r
\name{get.temperature}
\alias{get.temperature}
\title{Get temperature for current iteration}
\usage{
get.temperature(max.temp, max.iter, cur.iter)
}
\arguments{
\item{max.temp}{Max value of temperature to use (float)}

\item{max.iter}{Max number of iterations to search for optimal solution
(integer)}

\item{cur.iter}{Value of current iteration (integer)}
}
\value{
Value of temperture for the current iteration (integer)
}
\description{
Temperature decreases linearly with each iteration
}
\examples{
temp <- get.temperature(25, 100, 7)
}

