% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m3xlarge.runtimes.expdist}
\alias{m3xlarge.runtimes.expdist}
\title{Runtimes on m3xlarge instance}
\format{A data frame with 21750 rows and 2 variables
\describe{
 \item{size_bp}{size in bp}
 \item{runtime_sec}{runtime in seconds on m3xlarge instance}
}}
\source{
Generated via the R script /data-raw/generate_m3xlarge_runtimes_expdist.R in the source package
}
\usage{
m3xlarge.runtimes.expdist
}
\description{
Dataset containing exponentially distributed runtimes generated based on mean runtimes for jobs run on m3xlarge instances in EC2
}
\keyword{datasets}

