% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulated-annealing.R
\name{move.tasks}
\alias{move.tasks}
\title{Generate neighbor by moving 1 task}
\usage{
move.tasks(schedule, num.tasks, exchange = FALSE)
}
\arguments{
\item{schedule}{A list representing the schedule for which a neighbor is
desired}

\item{num.tasks}{Integer representing the number of tasks to be moved from 1
instance to another}

\item{exchange}{Exchange tasks between instances instead of moving them}
}
\value{
A list representing the neighboring schedule
}
\description{
Randomly select 2 instances in the cluster. Randomly select a task from one
of the instances and move it to the other instance. Simple random sampling
without replacement is used in both sampling stages.
}
\examples{
schedule <- get.initial.schedule(3, 1:30)
neighbor <- move.tasks(schedule, 1)
neighbor <- move.tasks(schedule, 1, exchange=TRUE)
}

