% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.r
\name{schedule}
\alias{schedule}
\title{Find optimal schedule}
\usage{
schedule(job, deadline, cluster.instance.type, cluster.size, max.iter, max.temp)
}
\value{
A list representing the optimal assignment that could be found under the given constraints
}
\description{
Want an assignment with >= 95% probability of completing job by the deadline with the lowest cost
}
\examples{
job <- seq(1:30)
deadline <- 3600
cluster.instance.type <- 'm3xlarge'
cluster.size <- 10
max.iter <- 10
max.temp <- 25
best.schedule <- schedule(job, deadline, cluster.instance.type, cluster.size, max.iter, max.temp)
}

