% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/schedulr.R
\docType{package}
\name{schedulr}
\alias{schedulr}
\alias{schedulr-package}
\title{schedulr: schedule jobs on machines}
\description{
schedulr provides functions to schedule jobs on machines based on type of job, type of machine and various input constraints
}
\details{
To learn more about schedulr, start with the vignettes:
\code{browseVignettes(package = "schedulr")}
}
\author{
Niranjan Vissa (niranjanv+schedulr@gmail.com)
}

