% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.R
\name{setup.trainingset.runtimes}
\alias{setup.trainingset.runtimes}
\title{Setup runtimes for given instance type}
\usage{
setup.trainingset.runtimes(instance.type, runtimes)
}
\arguments{
\item{instance.type}{Instance type of cluster (string).
All instances in the cluster are assumed to be of the same type}

\item{runtimes}{Matrix of runtimes for the given instance type.
Each row in the matrix represents a single training sample and has 2 columns.
The size column is the size of task that was processed.
The runtime_sec column is the time taken to process the task in seconds.}
}
\description{
All instances in a cluster are assumed to be of the same type
}
\examples{
runtimes <- cbind(rep(c(1,2), each=5), c(rpois(5,5), rpois(5,10)))
setup.trainingset.runtimes('m3xlarge', runtimes)
}

