% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.R
\name{compare.assignments}
\alias{compare.assignments}
\title{Compare 2 assignments based on their score}
\usage{
compare.assignments(cur.assignment, proposed.assignment, runtimes,
  runtimes.summary, deadline, max.temp, max.iter, cur.iter)
}
\arguments{
\item{cur.assignment}{Current assigment with score attribute (list)}

\item{proposed.assignment}{Proposed assignment with no score (list)}

\item{runtimes}{Matrix of runtimes for the given instance type. Each row in
the matrix represents a single training sample and has 2 columns. The size
column is the size of task that was processed. The runtime_sec column is the
time taken to process the task in seconds.}

\item{runtimes.summary}{Numeric matrix containing mean and variance of
runtimes for each size}

\item{deadline}{Time by which job must be complete (float). Same time units
as runtimes}

\item{max.temp}{Max temperature to use in the simulated annealing process
'(integer)}

\item{max.iter}{Max # iterations to use to find the optimal assignment via
simulated annealing (integer)}

\item{cur.iter}{Value of current iteration (integer)}
}
\value{
A list containing the accepted assignment and score
}
\description{
Scores are calculated for both assignments. If the score of the proposed
assignment is lower than the score for the current assignment, the proposed
assignment and score are returned. If the score of the proposed assignment is
greater than or equal to the current assignment, the a function of the
current temperature and the 2 scores is used to determine which assignment
to return.
}

