% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.r
\name{get.neighbor}
\alias{get.neighbor}
\title{Generate a neighbor to an assignment}
\usage{
get.neighbor(assignment)
}
\arguments{
\item{assignment}{A list representing a mapping of tasks to instances in a cluster}
}
\value{
A list representing the modified assignment of tasks to instances in the cluster
}
\description{
The input assignment is modified in one of several different ways, including
\itemize{
 \item Move a task from 1 instance to another
 \item Exchange a task with another instance
 \item Move 2 tasks from 1 instance to another
 \item Exchange 2 tasks with another instance
 \item Move 2 tasks from an instance to 2 other instance
 \item Exchange 2 tasks with 2 other instances
 \item and so on...
}
Only the first 2 methods are currently implemented with an equal probability of selecting either method.
}
\examples{
assignment <- get.initial.assignment(3, 1:30)
proposed.assignment <- get.neighbor(assignment)
}

