% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulated-annealing.r
\name{move.tasks}
\alias{move.tasks}
\title{Generate neighbor by moving 1 task}
\usage{
move.tasks(assignment, num.tasks, exchange = FALSE)
}
\arguments{
\item{assignment}{A list representing the assignment for which a neighbor is desired}

\item{num.tasks}{Integer representing the number of tasks to be moved from 1 instance to another}

\item{exchange}{Exchange tasks between instances instead of moving them}
}
\value{
A list representing the neighboring assignment
}
\description{
Randomly select 2 instances in the cluster. Randomly select a task from one of the instances and move it to the other instance. Simple random sampling without replacement is used in both sampling stages.
}
\examples{
assignment <- get.initial.assignment(3, 1:30)
neighbor <- move.tasks(assignment, 1)
neighbor <- move.tasks(assignment, 1, exchange=TRUE)
}

